Update script.js and index.html
Commit Date: 10/21/2025

Start stopwatch:

The stopwatch starts counting up as soon as the page is loaded. Adds to the imperativeness of the game.

Add Googly Eye effect:

The pupils within every eye on every card follow the mouse. This contributes to the overall  clean but fun aesthetic of the site.

Fix issues with VerifySet, RemoveSet, and related comparison functions:

Certain cases would be counted as a set even though they were not. These functions would compare the attributes of the first and second card, then the first and last card. There was one missing comparison between the second and third card, causing some invalid sets to be counted as valid ones. One more comparison was added to all of these functions.

Start game statistics:

There are three statistics added to make the game a bit more interesting. 1) a counter for how many cards have been taken out of the total 81 cards available. 2) An area for a pie chart to display the diversification of the kinds of sets found ( same color, same pattern, same, number, same shape, everything is different, etc…)

Add Replacement of cards taken by the player:

When a valid set is taken by the player, the cards within that set are removed from the board and replaced by a new one without changing the positions of the previous cards. This was added so that the player is able to play through a full game of set by themselves

Chore list:
	Update Rhombus striped pattern and Chevron outline pattern
	Add button to add up to 15 cards to the board
	Add Rule Card popup
	Add card animations
	Update Statistics
		Add pie chart display of kinds of sets taken
		Rename “sets found” to “cards taken”
	Update UI
		Green overlay for valid sets before they are taken
		Red overlay for invalid sets before the selection is reset
	Add game ending functions
		Quit button
		Stopwatch stops when game is ended
		Overall ending screen display, that includes the statistics from the round